add two tests
