+++ pgdrive/world/screen_shot_camera.py
